# packages/api-contract/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Stride API",
        "version": "0.1.0"
    },
    "paths": {
        "/inputs/{input_id}/logs": {
            "post": {
                "tags": [
                    "input-logs"
                ],
                "summary": "Create Input Log",
                "operationId": "create_input_log_inputs__input_id__logs_post",
                "parameters": [
                    {
                        "name": "input_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Input Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InputLogCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InputLogRead"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "input-logs"
                ],
                "summary": "Read Input Logs",
                "operationId": "read_input_logs_inputs__input_id__logs_get",
                "parameters": [
                    {
                        "name": "input_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Input Id"
                        }
                    },
                    {
                        "name": "range",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Range"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/InputLogRead"
                                    },
                                    "title": "Response Read Input Logs Inputs  Input Id  Logs Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/inputs/{input_id}/logs/{log_id}": {
            "delete": {
                "tags": [
                    "input-logs"
                ],
                "summary": "Delete Input Log",
                "operationId": "delete_input_log_inputs__input_id__logs__log_id__delete",
                "parameters": [
                    {
                        "name": "input_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Input Id"
                        }
                    },
                    {
                        "name": "log_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Log Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "InputLogCreate": {
                "properties": {
                    "value": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Value"
                    },
                    "attributes": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Attributes"
                    },
                    "occurred_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Occurred At"
                    }
                },
                "type": "object",
                "required": [
                    "occurred_at"
                ],
                "title": "InputLogCreate"
            },
            "InputLogRead": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "input_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Input Id"
                    },
                    "value": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Value"
                    },
                    "attributes": {
                        "anyOf": [
                            {
                                "additionalProperties": true,
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Attributes"
                    },
                    "occurred_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Occurred At"
                    },
                    "logged_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Logged At"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "input_id",
                    "value",
                    "attributes",
                    "occurred_at",
                    "logged_at"
                ],
                "title": "InputLogRead"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}